Refactor Discourse objects to be stateless
API clients are no longer attached to the various classes. Instead, the
client object must be passed as first argument to every method that
queries the forum.

This makes for a cleaner codebase: The various API objects can now be
serialized/deserialized without worrying about the unrelated client that
queried them in the first place.