Revert changes to JsonObject
Moved back to updating attributes of the base class with the incoming JSON dict. This enables classes to be properly inspected with vars().

Changed repr back. It still fails eval(), but this is a closer step to a correct object representation than the proposed implementation.

Made client a required incoming argument.